+++ app/forms/review_form.py
@@ -5,3 +5,4 @@ from wtforms.validators import DataRequired, NumberRange
 class ReviewForm(FlaskForm):
     review = TextAreaField('Review', validators=[DataRequired()])
     rating = IntegerField('Rating', validators=[DataRequired()])
+    product_id = IntegerField('ProductId', validators=[DataRequired()])
