+++ react-app/src/components/ProductsPage/EditProduct /EditProduct.js
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { useHistory, useParams, Redirect } from "react-router-dom";
+import { useHistory, useParams } from "react-router-dom";
 import { editOneProduct } from "../../../store/products";
 
 const EditProduct = ({ onClose }) => {
